Update existing serializers
