+++ ballot/serializers/referendum.py
@@ -1,9 +1,9 @@
+from rest_framework import serializers
+
 from .. import models
-from _django_utils.serializers import ExtendedModelSerializer
 
 
-class ReferendumSerializer(ExtendedModelSerializer):
+class ReferendumSerializer(serializers.ModelSerializer):
     class Meta:
         model = models.Referendum
-        exclude = ('contest_type',)
-        rename = dict(ballot='ballot_id')
+        exclude = ('ballot',)
